+++ MatrixSDK/VoIP/MXCall.m
@@ -294,7 +294,7 @@ NSString *const kMXCallSupportsTransferringStatusDidChange = @"kMXCallSupportsTr
                     },
                     @"version": kMXCallVersion,
                     @"lifetime": @(self->callManager.inviteLifetime),
-                    @"capabilities": @{@"m.call.transferee": @(YES)},
+                    @"capabilities": @{@"m.call.transferee": @(NO)},
                     @"party_id": self.partyId
                 } mutableCopy];
                 
@@ -370,7 +370,7 @@ NSString *const kMXCallSupportsTransferringStatusDidChange = @"kMXCallSupportsTr
                                                       @"type": kMXCallSessionDescriptionTypeStringAnswer,
                                                       @"sdp": sdpAnswer
                                                       },
-                                              @"capabilities": @{@"m.call.transferee": @(YES)},
+                                              @"capabilities": @{@"m.call.transferee": @(NO)},
                                               @"version": kMXCallVersion,
                                               @"party_id": self.partyId
                                               };
@@ -589,10 +589,14 @@ NSString *const kMXCallSupportsTransferringStatusDidChange = @"kMXCallSupportsTr
 
 - (BOOL)supportsTransferring
 {
-    if (callInviteEventContent && _selectedAnswer)
+    if (self.isIncoming)
+    {
+        return callInviteEventContent.capabilities.transferee;
+    }
+    else if (_selectedAnswer)
     {
         MXCallAnswerEventContent *content = [MXCallAnswerEventContent modelFromJSON:_selectedAnswer.content];
-        return callInviteEventContent.capabilities.transferee && content.capabilities.transferee;
+        return content.capabilities.transferee;
     }
     return NO;
 }
@@ -870,6 +874,25 @@ NSString *const kMXCallSupportsTransferringStatusDidChange = @"kMXCallSupportsTr
         if ([_delegate respondsToSelector:@selector(callConsultingStatusDidChange:)])
         {
             [_delegate callConsultingStatusDidChange:self];
+        }
+    }
+}
+
+- (void)setAssertedIdentity:(MXAssertedIdentityModel *)assertedIdentity
+{
+    if (![_assertedIdentity isEqual:assertedIdentity])
+    {
+        _assertedIdentity = assertedIdentity;
+        
+        if (self.isEstablished && _state != MXCallStateEnded)
+        {
+            //  reset call connected date
+            callConnectedDate = [NSDate date];
+        }
+        
+        if ([_delegate respondsToSelector:@selector(callAssertedIdentityDidChange:)])
+        {
+            [_delegate callAssertedIdentityDidChange:self];
         }
     }
 }
@@ -1295,12 +1318,12 @@ NSString *const kMXCallSupportsTransferringStatusDidChange = @"kMXCallSupportsTr
 
 - (void)handleCallReplaces:(MXEvent *)event
 {
-    //  TODO: Implement
+    
 }
 
 - (void)handleCallRejectReplacement:(MXEvent *)event
 {
-    //  TODO: Implement
+    
 }
 
 #pragma mark - Private methods