+++ CHANGES.rst
@@ -5,16 +5,42 @@ Changes to be released in next version
  * 
 
 🙌 Improvements
+ * 
+
+🐛 Bugfix
+ * 
+
+⚠️ API Changes
+ * 
+
+🗣 Translations
+ * 
+    
+🧱 Build
+ * 
+
+Others
+ * 
+
+Changes in 0.18.11 (2021-05-07)
+=================================================
+
+✨ Features
+ * 
+
+🙌 Improvements
  * MXCallKitAdapter: Update incoming calls if answered from application UI.
  * MXFileStore: Logs all files when a data corruption is detected (to track vector-im/element-ios/issues/4921).
  * MXCallManager: Fix call transfers flow for all types of transfers.
+ * VoIP: Implement asserted identity for calls: MSC3086 (matrix-org/matrix-doc/pull/3086).
 
 🐛 Bugfix
  * MXTools: Fix bad linkification of matrix alias and URL (vector-im/element-ios/issues/4258).
  * MXRoomSummary: Fix roomType property deserialization issue.
+ * MXCall: Disable call transferee capability & fix call transfer feature check.
 
 ⚠️ API Changes
- * 
+ * Spaces and room type: Remove all MSC1772 JSON key prefixes and use stable ones.
 
 🗣 Translations
  * 
@@ -28,6 +54,9 @@ Changes to be released in next version
 
 Others
  * 
+
+Improvements:
+
 
 Changes in 0.18.10 (2021-04-22)
 =================================================